- adding php 7.2 support
- fixing build issues in php 5.6 and php latest
- deprecating php 5.4
- fixing travis issue for new containers

+++ files/shared/scripts/install-git-crypt.sh
@@ -5,9 +5,14 @@ set -o pipefail # Return exit status of the last command in the pipe that failed
 
 readonly GIT_CRYPT_VERSION=0.5.0-2
 
-curl -L "https://github.com/AGWA/git-crypt/archive/debian/$GIT_CRYPT_VERSION.tar.gz" | tar zxv -C /var/tmp && \
-cd "/var/tmp/git-crypt-debian-$GIT_CRYPT_VERSION"
-make
-make install PREFIX=/usr/local
-cd -
-rm -f -r /var/tmp/git-crypt-debian-$GIT_CRYPT_VERSION
+source  /etc/os-release
+if [ ${VERSION:3:-1} == 'stretch' ] ; then
+        apt-get -y install git-crypt || true
+else
+  curl -L "https://github.com/AGWA/git-crypt/archive/debian/$GIT_CRYPT_VERSION.tar.gz" | tar zxv -C /var/tmp && \
+  cd "/var/tmp/git-crypt-debian-$GIT_CRYPT_VERSION"
+  make
+  make install PREFIX=/usr/local
+  cd -
+  rm -f -r /var/tmp/git-crypt-debian-$GIT_CRYPT_VERSION
+fi
